code/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
click==8.1.7
dnspython==2.6.1
fastapi==0.112.0
h11==0.14.0
idna==3.7
motor==3.5.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
wsproto==1.2.0
